+++ src/Operator/V1/CampaignOperator.php
@@ -9,10 +9,15 @@ use Dsl\MyTarget\Domain\V1\Enum\Status;
 use Dsl\MyTarget\Mapper\Mapper;
 use Dsl\MyTarget\Operator\V1\Fields\CampaignFields;
 use Dsl\MyTarget\Client;
+use Dsl\MyTarget\Context;
 use Dsl\MyTarget as f;
 
 class CampaignOperator
 {
+    const LIMIT_CREATE = "campaign-create";
+    const LIMIT_EDIT = "campaign-edit";
+    const LIMIT_FIND = "campaigns-find";
+
     /**
      * @var Client
      */
@@ -30,22 +35,14 @@ class CampaignOperator
     }
 
     /**
-     * @param string $username
-     * @return ClientCampaignOperator
-     */
-    public function forClient($username)
-    {
-        return new ClientCampaignOperator($username, $this->client, $this->mapper);
-    }
-
-    /**
      * @param MutateCampaign $campaign
-     * @param array|null $context
+     * @param Context|null $context
      *
      * @return Campaign
      */
-    public function create(MutateCampaign $campaign, array $context = null)
+    public function create(MutateCampaign $campaign, Context $context = null)
     {
+        $context = Context::withLimitBy($context, self::LIMIT_CREATE);
         $rawCampaign = $this->mapper->snapshot($campaign);
 
         $json = $this->client->post("/api/v1/campaigns.json", null, $rawCampaign, $context);
@@ -56,12 +53,26 @@ class CampaignOperator
     /**
      * @param int $id
      * @param MutateCampaign $campaign
-     * @param array|null $context
+     * @param Context|null $context
+     * @deprecated Use edit() instead. All editing actions will be consistently named across all Operators as edit*
      *
      * @return Campaign
      */
-    public function update($id, MutateCampaign $campaign, array $context = null)
+    public function update($id, MutateCampaign $campaign, Context $context = null)
     {
+        return $this->edit($id, $campaign, $context);
+    }
+
+    /**
+     * @param int $id
+     * @param MutateCampaign $campaign
+     * @param Context|null $context
+     *
+     * @return Campaign
+     */
+    public function edit($id, MutateCampaign $campaign, Context $context = null)
+    {
+        $context = Context::withLimitBy($context, self::LIMIT_EDIT);
         $rawCampaign = $this->mapper->snapshot($campaign);
 
         $json = $this->client->post(sprintf("/api/v1/campaigns/%d.json", $id), null, $rawCampaign, $context);
@@ -74,12 +85,13 @@ class CampaignOperator
      *
      * @param CampaignFields|null $fields
      * @param Status[]|null $withStatuses
-     * @param array|null $context
+     * @param Context|null $context
      *
      * @return CampaignStat[]
      */
-    public function all(CampaignFields $fields = null, array $withStatuses = null, array $context = null)
+    public function all(CampaignFields $fields = null, array $withStatuses = null, Context $context = null)
     {
+        $context = Context::withLimitBy($context, self::LIMIT_FIND);
         $fields = $fields ?: CampaignFields::create();
 
         $query = ["fields" => $this->mapFields($fields->getFields())];
@@ -91,8 +103,6 @@ class CampaignOperator
         if ($fields->hasField(CampaignFields::FIELD_BANNERS)) {
             $query["with_banners"] = "1";
         }
-
-        $context = (array)$context + ["limit-by" => "campaigns-all"];
 
         $json = $this->client->get("/api/v1/campaigns.json", $query, $context);
 
@@ -108,11 +118,11 @@ class CampaignOperator
      *
      * @param int $id
      * @param CampaignFields|null $fields
-     * @param array|null $context
+     * @param Context|null $context
      *
      * @return CampaignStat|null
      */
-    public function find($id, CampaignFields $fields = null, array $context = null)
+    public function find($id, CampaignFields $fields = null, Context $context = null)
     {
         $campaigns = $this->findAll([$id], $fields, null, $context);
 
@@ -125,14 +135,15 @@ class CampaignOperator
      * @param int[] $ids
      * @param CampaignFields|null $fields
      * @param Status[]|null $withStatuses
-     * @param array|null $context
+     * @param Context|null $context
      *
      * @return CampaignStat[]
      */
-    public function findAll(array $ids, CampaignFields $fields = null, array $withStatuses = null, array $context = null)
+    public function findAll(array $ids, CampaignFields $fields = null, array $withStatuses = null, Context $context = null)
     {
-        $context = (array)$context + ["limit-by" => "campaigns-find"];
+        $context = Context::withLimitBy($context, self::LIMIT_FIND);
 
+        $fields = $fields ?: CampaignFields::create();
         $query = ["fields" => $this->mapFields($fields->getFields())];
         if ($fields->hasField(CampaignFields::FIELD_BANNERS)) {
             $query["with_banners"] = "1";