+++ src/Operator/V1/PackagesOperator.php
@@ -3,11 +3,14 @@
 namespace Dsl\MyTarget\Operator\V1;
 
 use Dsl\MyTarget\Client;
+use Dsl\MyTarget\Context;
 use Dsl\MyTarget\Domain\V1\Campaign\Package;
 use Dsl\MyTarget\Mapper\Mapper;
 
 class PackagesOperator
 {
+    const LIMIT_FIND = "packages-find";
+
     /**
      * @var Client
      */
@@ -25,11 +28,12 @@ class PackagesOperator
     }
 
     /**
-     * @param array|null $context
+     * @param Context|null $context
      * @return Package[]
      */
-    public function all(array $context = null)
+    public function all(Context $context = null)
     {
+        $context = Context::withLimitBy($context, self::LIMIT_FIND);
         $json = $this->client->get('/api/v1/packages.json', null, $context);
 
         return array_map(function ($json) {