+++ src/Operator/V1/ProjectionOperator.php
@@ -3,12 +3,15 @@
 namespace Dsl\MyTarget\Operator\V1;
 
 use Dsl\MyTarget\Client;
+use Dsl\MyTarget\Context;
 use Dsl\MyTarget\Domain\V1\Campaign\Projection\Projection;
 use Dsl\MyTarget\Domain\V1\Campaign\Projection\ProjectionCampaign;
 use Dsl\MyTarget\Mapper\Mapper;
 
 class ProjectionOperator
 {
+    const LIMIT_PROJECTION = "projection";
+
     /**
      * @var Client
      */
@@ -27,12 +30,12 @@ class ProjectionOperator
 
     /**
      * @param ProjectionCampaign $campaign
-     * @param array|null $context
+     * @param Context|null $context
      * @return Projection
      */
-    public function projection(ProjectionCampaign $campaign, array $context = null)
+    public function projection(ProjectionCampaign $campaign, Context $context = null)
     {
-        $context = (array)$context + ["limit-by" => "projection"];
+        $context = Context::withLimitBy($context, self::LIMIT_PROJECTION);
         $data = $this->mapper->snapshot($campaign);
 
         $json = $this->client->post("/api/v1/projection.json", null, $data, $context);