+++ src/Operator/V1/Remarketing/RemarketingAppsOperator.php
@@ -3,11 +3,16 @@
 namespace Dsl\MyTarget\Operator\V1\Remarketing;
 
 use Dsl\MyTarget\Client;
+use Dsl\MyTarget\Context;
 use Dsl\MyTarget\Domain\V1\Remarketing\RemarketingApp;
 use Dsl\MyTarget\Mapper\Mapper;
 
 class RemarketingAppsOperator
 {
+    const LIMIT_CREATE = "remarketing-apps-create";
+    const LIMIT_FIND = "remarketing-apps-find";
+    const LIMIT_DELETE = "remarketing-apps-delete";
+
     /**
      * @var Client
      */
@@ -26,12 +31,13 @@ class RemarketingAppsOperator
 
     /**
      * @param RemarketingApp $app
-     * @param array|null $context
+     * @param Context|null $context
      *
      * @return RemarketingApp
      */
-    public function create(RemarketingApp $app, array $context = null)
+    public function create(RemarketingApp $app, Context $context = null)
     {
+        $context = Context::withLimitBy($context, self::LIMIT_CREATE);
         $rawApp = $this->mapper->snapshot($app);
 
         $json = $this->client->post("/api/v1/remarketing_apps.json", null, $rawApp, $context);
@@ -40,12 +46,13 @@ class RemarketingAppsOperator
     }
 
     /**
-     * @param array|null $context
+     * @param Context|null $context
      *
      * @return RemarketingApp[]
      */
-    public function all(array $context = null)
+    public function all(Context $context = null)
     {
+        $context = Context::withLimitBy($context, self::LIMIT_FIND);
         $json = $this->client->get("/api/v1/remarketing_apps.json", null, $context);
 
         return array_map(function ($json) {
@@ -55,11 +62,11 @@ class RemarketingAppsOperator
 
     /**
      * @param int $id
-     * @param array|null $context
+     * @param Context|null $context
      */
-    public function delete($id, array $context = null)
+    public function delete($id, Context $context = null)
     {
         $path = sprintf("/api/v1/remarketing_apps/%d.json", $id);
-        $this->client->delete($path, null, $context);
+        $this->client->delete($path, null, Context::withLimitBy($context, self::LIMIT_DELETE));
     }
 }