+++ src/Operator/V2/CreativeOperator.php
@@ -5,13 +5,16 @@ namespace Dsl\MyTarget\Operator\V2;
 use Dsl\MyTarget\Client;
 use Dsl\MyTarget\Domain\V2\Enum\CreativeType;
 use Dsl\MyTarget\Mapper\Mapper;
-use Dsl\MyTarget\Operator\Exception\UnexpectedFileArgumentException;
 use Dsl\MyTarget\Domain\V2\Creative;
 use Dsl\MyTarget\Domain\V2\UploadCreative;
 use Psr\Http\Message\StreamInterface;
+use Dsl\MyTarget as f;
+use Dsl\MyTarget\Context;
 
 class CreativeOperator
 {
+    const LIMIT_CREATE = "v2-creative-create";
+
     /**
      * @var Client
      */
@@ -32,27 +35,20 @@ class CreativeOperator
      * @param resource|string|StreamInterface $file Can be a StreamInterface instance, resource or a file path
      * @param CreativeType $type
      * @param UploadCreative $creative
-     * @param array|null $context
-     * @param str|null $filename
+     * @param string|null $filename
+     * @param Context|null $context
      *
      * @return Creative
      */
-    public function create($file, CreativeType $type, UploadCreative $creative, array $context = null, $filename = null)
+    public function create($file, CreativeType $type, UploadCreative $creative, $filename = null, Context $context = null)
     {
-        if (is_string($file)) { // assume it's a file path
-            $file = fopen($file, 'r');
-        }
-        if ( ! $file instanceof StreamInterface && ! is_resource($file)) {
-            throw new UnexpectedFileArgumentException($file);
-        }
+        $context = Context::withLimitBy($context, self::LIMIT_CREATE);
+        $file = f\streamOrResource($file);
 
         $rawCreative = $this->mapper->snapshot($creative);
         $body = [
-            ["name" => "file", "contents" => $file],
+            ["name" => "file", "contents" => $file, "filename" => $filename],
             ["name" => "data", "contents" => \json_encode($rawCreative)]];
-        if (null !== $filename) {
-            $body[0]['filename'] = $filename;
-        }
 
         $path = sprintf("/api/v2/content/%s.json", $type->getValue());
         $json = $this->client->postMultipart($path, $body, null, $context);