+++ src/Token/ClientCredentials/CredentialsProvider.php
@@ -2,15 +2,16 @@
 
 namespace Dsl\MyTarget\Token\ClientCredentials;
 
+use Dsl\MyTarget\Context;
 use Psr\Http\Message\RequestInterface;
 
 interface CredentialsProvider
 {
     /**
      * @param RequestInterface $request
-     * @param array|null $context
+     * @param Context $context
      *
      * @return Credentials
      */
-    public function getCredentials(RequestInterface $request, array $context = null);
+    public function getCredentials(RequestInterface $request, Context $context);
 }