+++ src/Token/DoctrineCacheTokenStorage.php
@@ -5,6 +5,7 @@ namespace Dsl\MyTarget\Token;
 use Doctrine\Common\Cache\Cache;
 use Psr\Http\Message\RequestInterface;
 use Dsl\MyTarget as f;
+use Dsl\MyTarget\Context;
 
 /**
  * Token storage implementation that depends on "doctrine/cache" composer package
@@ -23,7 +24,7 @@ class DoctrineCacheTokenStorage implements TokenStorage
 
     /**
      * @param Cache $cache
-     * @param callable $hashFunction callable(string $id, RequestInterface, $context): string Identity function used as a default
+     * @param callable $hashFunction callable(string $id, RequestInterface, Context $context): string Identity function used as a default
      */
     public function __construct(Cache $cache, callable $hashFunction = null)
     {
@@ -34,7 +35,7 @@ class DoctrineCacheTokenStorage implements TokenStorage
     /**
      * @inheritdoc
      */
-    public function getToken($id, RequestInterface $request, array $context = null)
+    public function getToken($id, RequestInterface $request, Context $context)
     {
         $id = call_user_func($this->hashFunction, $id, $request, $context);
         $tokenArray = $this->cache->fetch($id);
@@ -49,7 +50,7 @@ class DoctrineCacheTokenStorage implements TokenStorage
     /**
      * @inheritdoc
      */
-    public function updateToken($id, Token $token, RequestInterface $request, array $context = null)
+    public function updateToken($id, Token $token, RequestInterface $request, Context $context)
     {
         $id = call_user_func($this->hashFunction, $id, $request, $context);