+++ src/Token/TokenManager.php
@@ -2,12 +2,15 @@
 
 namespace Dsl\MyTarget\Token;
 
+use Dsl\MyTarget\Context;
 use Dsl\MyTarget\Token\ClientCredentials\CredentialsProvider;
 use Dsl\MyTarget\Token\Exception\TokenDeletedException;
 use Dsl\MyTarget\Token\Exception\TokenLimitReachedException;
 use Dsl\MyTarget\Token\Exception\TokenLockException;
 use Dsl\MyTarget\Token\Exception\TokenRequestException;
 use Psr\Http\Message\RequestInterface;
+use Psr\Log\LoggerInterface;
+use Psr\Log\NullLogger;
 
 class TokenManager
 {
@@ -32,6 +35,11 @@ class TokenManager
     private $lockManager;
 
     /**
+     * @var LoggerInterface
+     */
+    private $logger;
+
+    /**
      * @var callable callable(): \DateTimeInterface Returns current moment
      */
     private $momentGenerator;
@@ -51,6 +59,7 @@ class TokenManager
         $this->momentGenerator = function () {
             return new \DateTimeImmutable();
         };
+        $this->logger = new NullLogger();
     }
 
     /**
@@ -61,41 +70,14 @@ class TokenManager
         $this->momentGenerator = $cb;
     }
 
-    /**
-     * @param RequestInterface $request
-     * @param string           $username
-     * @param array|null       $context
-     *
-     * @return Token|null
-     *
-     * @throws TokenLockException
-     * @throws TokenRequestException
-     * @throws TokenLimitReachedException
-     */
-    public function getClientToken(RequestInterface $request, $username, array $context = null)
+    public function setLogger(LoggerInterface $logger)
     {
-        return $this->doGetToken($request, $username, $context);
+        $this->logger = $logger;
     }
 
     /**
      * @param RequestInterface $request
-     * @param array|null $context
-     *
-     * @return Token|null
-     *
-     * @throws TokenLockException
-     * @throws TokenRequestException
-     * @throws TokenLimitReachedException
-     */
-    public function getToken(RequestInterface $request, array $context = null)
-    {
-        return $this->doGetToken($request, null, $context);
-    }
-
-    /**
-     * @param RequestInterface $request
-     * @param string|null      $username
-     * @param array|null       $context
+     * @param Context $context
      *
      * @return Token|null
      *
@@ -104,18 +86,31 @@ class TokenManager
      * @throws TokenLimitReachedException
      * @throws \Exception
      */
-    private function doGetToken(RequestInterface $request, $username = null, array $context = null)
+    public function getToken(RequestInterface $request, Context $context)
     {
         $credentials = $this->credentials->getCredentials($request, $context);
-        $id = $username ?: $credentials->getClientId();
+        $id = $context->hasUsername() ? $context->getUsername() : $credentials->getClientId();
 
         $now = call_user_func($this->momentGenerator);
         $token = $this->storage->getToken($id, $request, $context);
 
         if ( ! $token || $token->isExpiredAt($now)) {
 
+            if ($token) {
+                $this->logger->debug("Token({$id}) was found but is expired, will refresh it");
+            } else {
+                $this->logger->debug("Token({$id}) was not found, will acquire it");
+            }
+
             if ($this->lockManager) {
-                $this->lockManager->lock($id);
+                $this->logger->debug("Token({$id}) trying to get a lock");
+                try {
+                    $this->lockManager->lock($id);
+                    $this->logger->debug("Token({$id}) locked");
+                } catch (TokenLockException $e) {
+                    $this->logger->debug("Token({$id}) couldn't get a lock");
+                    throw $e;
+                }
             }
 
             try {
@@ -123,26 +118,36 @@ class TokenManager
                     try {
                         $token = $this->acquirer->refresh($request, $now, $token->getRefreshToken(), $context);
                         $this->storage->updateToken($id, $token, $request, $context);
+                        $this->logger->debug("Token({$id}) refreshed the token and put it in the storage");
                     } catch (TokenDeletedException $e) {
+                        $this->logger->debug("Token({$id}) was deleted, will try to acquire a new one");
+
                         // 30 days token expire, we should get new token
                         $token = null;
                     }
                 }
 
                 if ( ! $token) {
-                    $token = $this->acquirer->acquire($request, $now, $username, $context);
+                    $token = $this->acquirer->acquire($request, $now, $context);
                     $this->storage->updateToken($id, $token, $request, $context);
+                    $this->logger->debug("Token({$id}) was acquired and put in the storage");
                 }
 
                 if ($this->lockManager) {
                     $this->lockManager->unlock($id);
+                    $this->logger->debug("Token({$id}) unlocked");
                 }
             } catch (TokenLimitReachedException $e) {
+                $this->logger->debug("Token({$id}) limit reached");
                 throw $e;
             } catch (\Exception $e) {
+                $exClass = get_class($e);
+                $this->logger->debug("Token({$id}) exception {$exClass} with message: {$e->getMessage()}");
+
                 // finally works incorrectly with Redis in php5.5
                 if ($this->lockManager) {
                     $this->lockManager->unlock($id);
+                    $this->logger->debug("Token({$id}) unlocked");
                 }
 
                 throw $e;
@@ -153,18 +158,19 @@ class TokenManager
     }
 
     /**
-     * @param Token            $token
+     * @param Token $token
      * @param RequestInterface $request
-     * @param string|null      $account
-     * @param string|null      $username
-     * @param array|null       $context
+     * @param string|null $account
+     * @param string|null $username
+     * @param Context $context
      */
-    public function expireToken(Token $token, RequestInterface $request, $account = null, $username = null, array $context = null)
+    public function expireToken(Token $token, RequestInterface $request, $account = null, $username = null, Context $context)
     {
         $moment = call_user_func($this->momentGenerator);
 
         $expired = new Token($token->getAccessToken(), $token->getTokenType(), $moment, $token->getRefreshToken());
         $id = $username ?: $account;
         $this->storage->updateToken($id, $expired, $request, $context);
+        $this->logger->debug("Token({$id}) forcibly expired the token");
     }
 }