+++ src/Token/TokenStorage.php
@@ -2,6 +2,7 @@
 
 namespace Dsl\MyTarget\Token;
 
+use Dsl\MyTarget\Context;
 use Psr\Http\Message\RequestInterface;
 
 interface TokenStorage
@@ -9,17 +10,17 @@ interface TokenStorage
     /**
      * @param string $id
      * @param RequestInterface $request
-     * @param array|null $context
+     * @param Context $context
      *
      * @return Token
      */
-    public function getToken($id,  RequestInterface $request, array $context = null);
+    public function getToken($id,  RequestInterface $request, Context $context);
 
     /**
      * @param string $id
      * @param Token $token
      * @param RequestInterface $request
-     * @param array|null $context
+     * @param Context $context
      */
-    public function updateToken($id, Token $token, RequestInterface $request, array $context = null);
+    public function updateToken($id, Token $token, RequestInterface $request, Context $context);
 }