+++ src/Transport/HttpTransport.php
@@ -2,6 +2,7 @@
 
 namespace Dsl\MyTarget\Transport;
 
+use Dsl\MyTarget\Context;
 use Dsl\MyTarget\Transport\Exception\NetworkException;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
@@ -10,10 +11,10 @@ interface HttpTransport
 {
     /**
      * @param RequestInterface $request
-     * @param array|null $context
+     * @param Context $context
      *
      * @return ResponseInterface
      * @throws NetworkException
      */
-    public function request(RequestInterface $request, array $context = null);
+    public function request(RequestInterface $request, Context $context);
 }